fix: correctly identify empty repository when commiting
The `is_empty` function exposed by libgit2 requires the default branch to be
"master". If a repository is created with a default branch not named master `is_empty`
will return false even when the repository is empty.

The fix is simply to try and access the HEAD of the repository. If the head
cant be accessed it means that the repository is still empty.

Closes #145