fix: handle grafted commit for shallow clone repository
Shallow clone is not supported yet in libgit2 (https://github.com/rust-lang/git2-rs/issues/796).
The starting (grafted) commit oid of a shallow clone repository can be seen but is not accessible.
A quick fix for this is to check weither the starting commit is reachable and fallback to the first
reachable one if not.

closes #173