feat(cli): generate man pages with clap-mangen
To generate man pages with clap_mangen the Cli struct needs to be accessible from the crate public api.
Because of this the clap app has moved to the crate library while the code responsible for handling the 
user inputs stays in the cog.rs bin file. It might be a good idea to split cocogitto into several crates
and use a cargo workspace instead of a single crate with multiple binaries.

closes #11