feat(bump): support annotated tags
Many projects use annotated commits for their releases. This commit
extends `cog bump` with the `-A MSG`/`--annoted MSG` argument, which
will create an annotated tag with the tag message `MSG`.

Much like the bump hooks, `{{version}}` and `{{latest}}` can be used in
the provided tag message and will be replaced accordingly.