feat(check): Support commit scope checking
Allowed scopes can now be defined in "cog.toml" with the top-level
"scopes" keyword. It takes an array of strings which are the scopes that
are allowed to be used.

If the "scopes" keyword isn't given, all scopes are allowed. If an empty
array of scopes is given, no scopes are allowed at all.

References: https://github.com/cocogitto/cocogitto/issues/42
Signed-off-by: Thomas Böhler <witcher@wiredspace.de>